makaogift/src/main/java/kr/megaptera/makaogift/services/UserService.java
package kr.megaptera.makaogift.services;

import kr.megaptera.makaogift.exceptions.UserNotFound;
import kr.megaptera.makaogift.models.User;
import kr.megaptera.makaogift.repositories.UserRepository;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;

@Service
@Transactional
public class UserService {
    private final UserRepository userRepository;

    public UserService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    public Long amount(String userName) {
        User user = userRepository.findByUserName(userName)
                .orElseThrow(() -> new UserNotFound());
        return user.amount();
    }
}
